day1/requirements.txt
litellm==1.44.0
python-dotenv==1.0.1
openai>=1.0.0